refactor(website): split TasksVisual into a folder structure
TasksVisual.tsx shell + barrel, with TaskRow in components/, the Task type
in types/, and TASKS / derived TASK_GROUPS in data/. RecordTabHeader stays
shared at the product-feature level.
