+++ packages/twenty-website/src/sections/product-feature/TasksVisual/index.ts
@@ -1,0 +1,1 @@
+export { TasksVisual } from './TasksVisual';
