+++ packages/twenty-website/src/sections/product-feature/TasksVisual/types/task.ts
@@ -1,0 +1,12 @@
+type TaskTarget = {
+  avatarUrl: string;
+  name: string;
+};
+
+export type Task = {
+  body: string;
+  done: boolean;
+  due: string;
+  target: TaskTarget;
+  title: string;
+};
